adds confirmation box before proceeding with deletion

+++ telli-react/src/components/DeleteWatchListButton.tsx
@@ -1,32 +1,38 @@
 interface DeleteWatchListButtonProps {
-    watchListId: number;
-    onDelete: () => void;
+  watchListId: number;
+  onDelete: () => void;
 }
 
-export default function DeleteWatchListButton({ watchListId, onDelete }: DeleteWatchListButtonProps) {
-
-    async function handleDelete() {
-        try {
-            const response = await fetch(`http://localhost:8080/api/delete-watchlist/${watchListId}`, {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-            });
-            if (response.ok) {
-                // POssibly add alert box here before proceeding with deletion?
-                onDelete();
-            } else {
-                console.error("Error deleting existing WatchList.")
-            }
-        } catch (error) {
-            console.error("Error deleting existing WatchList: ", error);
+export default function DeleteWatchListButton({
+  watchListId,
+  onDelete,
+}: DeleteWatchListButtonProps) {
+  async function handleDelete() {
+    if (confirm("Your Watch List will be deleted permanently!")) {
+      try {
+        const response = await fetch(
+          `http://localhost:8080/api/delete-watchlist/${watchListId}`,
+          {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+          }
+        );
+        if (response.ok) {
+          onDelete();
+        } else {
+          console.error("Error deleting existing WatchList.");
         }
+      } catch (error) {
+        console.error("Error deleting existing WatchList: ", error);
+      }
     }
+  }
 
-    return (
-        <>
-            <button onClick={handleDelete}>Delete</button>
-        </>
-    )
+  return (
+    <>
+      <button onClick={handleDelete}>Delete</button>
+    </>
+  );
 }
